Refactor authentication components to use consistent button styles and improve code readability
