+++ client/app/(tabs)/profile/AccountScreen.tsx
@@ -1,0 +1,170 @@
+import { View, TouchableOpacity, StyleSheet, Text } from "react-native";
+import ClerkSettings, {
+  ClerkSettingsRefType,
+} from "@/app/(auth)/ClerkSettings";
+import { useUser } from "@clerk/clerk-expo";
+import IconArrowBack from "@/assets/icons/IconArrowBack";
+import { Logo } from "@/components/Logos";
+import { Gaps } from "@/styles/theme";
+import { useRouter } from "expo-router";
+import React, { useRef, useState, useEffect, useCallback } from "react";
+import { useFocusEffect } from "expo-router";
+import { useGlobalLoading } from "@/providers/GlobalLoadingProvider";
+import AsyncStorage from "@react-native-async-storage/async-storage";
+import Loading from "@/app/Loading";
+
+const AccountScreen = () => {
+  const router = useRouter();
+  const { isAuthenticated, refreshGlobalState, isGloballyLoading } =
+    useGlobalLoading();
+  const isMounted = useRef(true);
+  const [refreshKey, setRefreshKey] = useState(0);
+  const hasFocusedRef = useRef(false);
+  const clerkSettingsRef = useRef<ClerkSettingsRefType>(null);
+  const [passwordResetFlag, setPasswordResetFlag] = useState<string | null>(
+    null
+  );
+  const { user } = useUser();
+  // Check for password reset flag on mount and refresh
+  useEffect(() => {
+    let isMounted = true;
+
+    const checkPasswordResetFlag = async () => {
+      try {
+        const resetFlag = await AsyncStorage.getItem("password_recently_reset");
+        if (isMounted) {
+          setPasswordResetFlag(resetFlag);
+        }
+      } catch (err) {
+        console.log("Error checking password reset flag:", err);
+      }
+    };
+
+    checkPasswordResetFlag();
+    return () => {
+      isMounted = false;
+    };
+  }, [refreshKey]);
+
+  // IMPORTANT: Trigger manual refresh via ref when auth state changes
+  useEffect(() => {
+    if (!isMounted.current || !clerkSettingsRef.current) return;
+
+    const timer = setTimeout(() => {
+      if (isMounted.current && clerkSettingsRef.current?.manualRefresh) {
+        clerkSettingsRef.current.manualRefresh();
+      }
+    }, 500);
+
+    return () => clearTimeout(timer);
+  }, [refreshKey, passwordResetFlag, isAuthenticated]);
+
+  // Track mounted state and save last screen
+  useEffect(() => {
+    isMounted.current = true;
+    AsyncStorage.setItem("last_screen", "/(tabs)/profile");
+    return () => {
+      isMounted.current = false;
+    };
+  }, []);
+
+  // Store refreshGlobalState in a ref to avoid dependency loops
+  const refreshGlobalStateRef = useRef(refreshGlobalState);
+  useEffect(() => {
+    refreshGlobalStateRef.current = refreshGlobalState;
+  }, [refreshGlobalState]);
+
+  // IMPORTANT: Reset focus flag after timeout to allow future refreshes when tab is revisited
+  useEffect(() => {
+    if (hasFocusedRef.current) {
+      const timer = setTimeout(() => {
+        hasFocusedRef.current = false;
+      }, 5000); // Reset after 5 seconds to allow future focus events
+      return () => clearTimeout(timer);
+    }
+  });
+
+  // CRITICAL: Handle tab focus and password reset detection
+  useFocusEffect(
+    useCallback(() => {
+      const timer = setTimeout(async () => {
+        if (hasFocusedRef.current) return;
+
+        console.log("Profile tab focused - refreshing auth state");
+        hasFocusedRef.current = true;
+
+        if (isMounted.current) {
+          try {
+            // IMPORTANT: Special handling for password reset flag
+            const resetFlag = await AsyncStorage.getItem(
+              "password_recently_reset"
+            );
+            if (resetFlag === "true") {
+              // Ensure flag persists for ClerkSettings to detect
+              await AsyncStorage.setItem(
+                "password_recently_reset_persist",
+                "true"
+              );
+
+              // Delay cleanup to ensure proper detection
+              setTimeout(() => {
+                AsyncStorage.removeItem("password_recently_reset").catch(
+                  (err) =>
+                    console.log("Error clearing password reset flag:", err)
+                );
+              }, 2000);
+            }
+
+            // Trigger refresh cascade
+            setRefreshKey((prev) => prev + 1);
+          } catch (err) {
+            console.log("Error in profile tab focus:", err);
+          }
+        }
+      }, 100);
+
+      return () => clearTimeout(timer);
+    }, [])
+  );
+
+  if (isGloballyLoading) {
+    return <Loading />;
+  }
+  return (
+    <View style={styles.container}>
+      <TouchableOpacity
+        style={styles.backButton}
+        onPress={() => router.back()}
+        accessibilityLabel="Go back"
+      >
+        <IconArrowBack />
+      </TouchableOpacity>
+      <View style={{ marginBottom: Gaps.g16 }}>
+        <Logo size="small" />
+      </View>
+
+      <View style={styles.buttonContainer}>
+        <ClerkSettings ref={clerkSettingsRef} refreshKey={refreshKey} />
+      </View>
+    </View>
+  );
+};
+export default AccountScreen;
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    marginTop: Gaps.g80,
+    alignItems: "center",
+  },
+  backButton: {
+    position: "absolute",
+    top: -8,
+    left: 16,
+    zIndex: 10,
+  },
+  buttonContainer: {
+    marginTop: Gaps.g8,
+    gap: Gaps.g16,
+  },
+});
